refactored logic into service class on backend
